deployments/liddy_voice/cerebrium/requirements-cerebrium.txt
# Core LiveKit dependencies
livekit==0.11.1
livekit-agents==0.8.0
livekit-plugins-openai==0.7.3
livekit-plugins-deepgram==0.6.4
livekit-plugins-elevenlabs==0.7.0
livekit-plugins-google==0.6.1
livekit-plugins-silero==0.7.0
livekit-plugins-assemblyai==0.2.8

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.21.0
google-generativeai
cerebras
langfuse>=2.0.0
tiktoken

# Vector search and data
pinecone-client>=3.0.0
redis>=5.0.0
numpy
pandas

# Web and async
aiohttp>=3.9.0
websockets>=12.0
httpx
pydantic>=2.0.0
python-dotenv

# Google Cloud
google-cloud-storage>=2.10.0
google-auth>=2.0.0
google-auth-oauthlib
google-auth-httplib2

# Audio processing
assemblyai
sounddevice
scipy

# Utilities
python-multipart
uvloop
psutil